refactor: remove dependency on data layer in view layer
