+++ app/src/main/java/woowacourse/movie/presenter/reservation/ReservationHistoryPresenter.kt
@@ -5,11 +5,12 @@ import woowacourse.movie.data.reservation.ReservationData
 import woowacourse.movie.domain.reservation.ReservationSortingPolicy
 import woowacourse.movie.domain.reservation.ShowtimeAscendingPolicy
 import woowacourse.movie.domain.ticket.Reservation
+import woowacourse.movie.view.reservation.LocalReservationData
 import kotlin.concurrent.thread
 
 class ReservationHistoryPresenter(
     private val view: ReservationHistoryContract.View,
-    private val reservationData: ReservationData,
+    private val reservationData: ReservationData = LocalReservationData,
     private val reservationSortingPolicy: ReservationSortingPolicy = ShowtimeAscendingPolicy(),
 ) : ReservationHistoryContract.Presenter {
     override fun fetchReservationHistories() {
